refactor: satisfy pedantic clippy (mostly)
I'm using the `ref` on the `field` argument in `try_from` to avoid
having to add additional lines. The trait necessitates that the function
take ownership of the value, but I'm unable to extract each field on the
`syn::Field` without running into a partial move...

For now, clippy will have to just deal with it.